move handleLike... to postUtils
